moved back to using the texture for sin/cos l/b
moved the multiplication by V back to the GPU, while on the CPU it uses a lot of memory because we have to save the stream_integral after every loop
of r_step.  This reduces the precision to 11 digits.  Ultimately the issue is because we do the calculations in the following loop on the GPU

for r
  for mu/nu

while the CPU does it like so

for mu
 for nu
   for r

resulting in small differences in the final answer